requirements.txt
aiohttp==3.9.5
datasets==2.20.0
fastapi==0.111.1
gradio==4.38.1
huggingface_hub==0.24.0
matplotlib==3.9.1
numpy==2.0.1
pandas==2.2.2
pydantic==2.8.2
python-dotenv==1.0.1
rouge_score==0.1.2
scipy==1.14.0
seaborn==0.13.2
setuptools==69.5.1
spacy==3.7.5
torch==2.3.1
transformers==4.42.4
trl==0.8.6
unsloth==2024.7
uvicorn==0.30.3
